+++ sources/apple/ios/AdMobWrapper/Ads/InterstitialAdWrapper.swift
@@ -31,7 +31,7 @@ import UIKit
             
             if let error = error {
                 let nsError = error as NSError
-                self.loadListener?.onAdFailedToLoad(errorCode: Int32(nsError.code),
+                self.loadListener?.onAdFailedToLoad(errorCode: nsError.code,
                                                     errorMessage: nsError.localizedDescription)
                 return
             }
@@ -77,7 +77,7 @@ extension InterstitialAdWrapper: FullScreenContentDelegate {
     public func ad(_ ad: FullScreenPresentingAd,
                    didFailToPresentFullScreenContentWithError error: Error) {
         let nsError = error as NSError
-        self.eventListener?.onAdFailedToShow(errorCode: Int32(nsError.code),
+        self.eventListener?.onAdFailedToShow(errorCode: nsError.code,
                                              errorMessage: nsError.localizedDescription)
         self.interstitialAd = nil
     }
